[process] For now we're keeping usec an integer, but maybe will go floating point one day. 'Far away there in the sunshine are my highest aspirations. I may not reach them, but I can look up and see their beauty, believe in them, and try to follow where they lead.' --Louisa May Alcott, <Little Women>

+++ data/process.py
@@ -22,7 +22,7 @@ class Process(object):
     def generate_occurence_usec(self, current_usec):
         # TODO...
         import random
-        self.occurence_usec = current_usec + (random.random() * 1e6)
+        self.occurence_usec = current_usec + int((random.random() * 1e6))
 
     def perform(self, step, lattice):
         if not self.is_still_performable(lattice):
